refactor: split whatsapp CLI god file into focused modules, add Go CI checks
Split the 2441-line whatsapp.go into 7 focused files (client, events,
messaging, contacts, groups, media, chat_ops). Extract all magic numbers
to constants.go with named constants. Remove dead code (unused SendMessage,
SendAudioMessageWithPresence, getChatNameFromConversation, MessageContext,
redundant error check, built-in min). Extract shared helpers (humanDelay,
randomDelay, writeNotificationFile). Unify ListGroups as ListChats
specialization. Refactor 500-line switch into named cmdXxx functions.
Fix archive-chat flag inconsistency (--jid → --to). Add Go format/vet
CI checks to pr-checks.yml.
